backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
pymysql
python-dotenv
flask-migrate
cryptography